requirements.txt
werkzeug
Flask==1.1.1
gunicorn==20.0.2
boto3==1.10.26
python-dotenv==0.10.3